Cleanup stuff from another branch
